Split models into different files
