+++ teccloud-api/package.json
@@ -30,7 +30,6 @@
     "@types/express": "^4.17.14",
     "@types/multer": "^1.4.7",
     "@types/node": "18.7.14",
-    "cross-env": "^7.0.3",
     "nodemon": "^2.0.20",
     "typescript": "4.8.2"
   }
