+++ teccloud-api/src/controllers/File/index.ts
@@ -13,8 +13,8 @@ const multerInstance = multer({
 const router = Router({ mergeParams: true });
 
 router.post('/upload', auth, FileController.upload(multerInstance));
-router.delete('/:fileId', auth, FileController.delete());
-router.get('/download/:fileId', auth, FileController.download());
+router.delete('/:fileName', auth, FileController.delete());
+router.get('/download/:fileName', auth, FileController.download());
 router.get('/:folderId', auth, FileController.get());
 
 export default router;
