+++ teccloud-api/src/db/fileaccess.ts
@@ -1,0 +1,53 @@
+import {
+  Model,
+  DataTypes,
+  CreationOptional,
+  InferAttributes,
+  InferCreationAttributes,
+} from 'sequelize';
+import { sequelize, File, User } from './index';
+
+export class FileAccess extends Model<
+  InferAttributes<FileAccess>,
+  InferCreationAttributes<FileAccess>
+> {
+  declare id: CreationOptional<number>;
+  declare fileId: number;
+  declare userId: number;
+  declare createdAt: CreationOptional<Date>;
+  declare updatedAt: CreationOptional<Date>;
+}
+
+FileAccess.init(
+  {
+    id: {
+      type: DataTypes.INTEGER,
+      autoIncrement: true,
+      primaryKey: true,
+    },
+    fileId: {
+      type: DataTypes.INTEGER,
+      references: {
+        model: File,
+        key: 'id',
+      },
+      allowNull: false,
+    },
+    userId: {
+      type: DataTypes.INTEGER,
+      references: {
+        model: User,
+        key: 'id',
+      },
+      allowNull: false,
+    },
+    createdAt: DataTypes.DATE,
+    updatedAt: DataTypes.DATE,
+  },
+  {
+    sequelize,
+    modelName: 'file_access',
+    timestamps: true,
+    indexes: [{ fields: ['fileId', 'userId'] }],
+  },
+);
