+++ teccloud-api/src/db/index.ts
@@ -5,4 +5,25 @@ export const sequelize = new Sequelize({
   storage: './teccloud.sqlite',
 });
 
-export * from './models';
+import { File } from './file';
+import { FileAccess } from './fileaccess';
+import { Folder } from './folder';
+import { Page } from './page';
+import { User } from './user';
+
+User.hasOne(Folder);
+Folder.belongsTo(User);
+
+Folder.belongsTo(Folder);
+Folder.hasMany(Folder);
+
+Folder.hasMany(File);
+File.belongsTo(Folder);
+
+User.belongsToMany(File, { through: FileAccess });
+File.belongsToMany(User, { through: FileAccess });
+
+File.hasMany(Page);
+Page.belongsTo(File);
+
+export { File, FileAccess, Folder, Page, User };
