+++ teccloud-api/src/db/user.ts
@@ -1,0 +1,76 @@
+import {
+  Model,
+  DataTypes,
+  CreationOptional,
+  InferAttributes,
+  InferCreationAttributes,
+} from 'sequelize';
+import * as jwt from 'jsonwebtoken';
+import bcrypt from 'bcrypt';
+import { sequelize, Folder } from './index';
+
+export class User extends Model<
+  InferAttributes<User>,
+  InferCreationAttributes<User>
+> {
+  declare id: CreationOptional<number>;
+  declare firstName: string;
+  declare lastName: string;
+  declare username: string;
+  declare password: string;
+  declare folderId: number;
+  declare token: CreationOptional<string>;
+  declare createdAt: CreationOptional<Date>;
+  declare updatedAt: CreationOptional<Date>;
+
+  async generateToken(): Promise<string> {
+    const jwtSecret = process.env.JWT_SECRET;
+    if (!jwtSecret) {
+      throw new Error('No JWT Secret has been defined');
+    }
+    const token = jwt.sign({ id: this.id.toString() }, jwtSecret, {
+      expiresIn: '10h',
+    });
+    this.token = token;
+    await this.save();
+    return token;
+  }
+
+  comparePassword(password: string): Promise<boolean> {
+    return bcrypt.compare(password, this.password);
+  }
+
+  static hashPassword(password: string): Promise<string> {
+    return bcrypt.hash(password, 10);
+  }
+}
+
+User.init(
+  {
+    id: {
+      type: DataTypes.INTEGER,
+      autoIncrement: true,
+      primaryKey: true,
+    },
+    firstName: { type: DataTypes.STRING, allowNull: false },
+    lastName: { type: DataTypes.STRING, allowNull: false },
+    username: { type: DataTypes.STRING, allowNull: false, unique: true },
+    password: { type: DataTypes.STRING, allowNull: false },
+    token: { type: DataTypes.STRING, allowNull: true },
+    folderId: {
+      type: DataTypes.INTEGER,
+      references: {
+        model: Folder,
+        key: 'id',
+      },
+      allowNull: true,
+    },
+    createdAt: DataTypes.DATE,
+    updatedAt: DataTypes.DATE,
+  },
+  {
+    sequelize,
+    modelName: 'users',
+    timestamps: true,
+  },
+);
