+++ teccloud-api/src/utils/files.ts
@@ -1,4 +1,4 @@
 import path from 'path';
 
-export const get_file_in_server_path = (fileId: string): string =>
-  path.resolve(path.join(process.env.FILES_FOLDER as string, fileId));
+export const get_file_server_path = (fileName: string): string =>
+  path.resolve(path.join(process.env.FILES_FOLDER as string, fileName));
