+++ teccloud-server/components/SingleFile.tsx
@@ -6,45 +6,43 @@ import Typography from '@mui/material/Typography';
 import DeleteFileModal from './DeleteFileModal';
 
 interface SingleFileProps {
-  fileId: string;
   fileName: string;
+  originalName: string;
 }
 
-const SingleFile = ({ fileId, fileName }: SingleFileProps) => {
-  return (
-    <Box
-      sx={{
-        height: '54px',
-        borderTop: 1,
-        borderBottom: 1,
-        borderColor: 'primary.main',
-        alignItems: 'center',
-        margin: '5px',
-      }}
+const SingleFile = ({ fileName, originalName }: SingleFileProps) => (
+  <Box
+    sx={{
+      height: '54px',
+      borderTop: 1,
+      borderBottom: 1,
+      borderColor: 'primary.main',
+      alignItems: 'center',
+      margin: '5px',
+    }}
+  >
+    <Stack
+      direction='row'
+      justifyContent='flex-start'
+      alignItems='center'
+      spacing={1}
     >
-      <Stack
-        direction='row'
-        justifyContent='flex-start'
-        alignItems='center'
-        spacing={1}
+      <IconButton
+        aria-label='delete'
+        size='large'
+        target='_blank'
+        href={`http://localhost:3001/files/download/${fileName}`}
       >
-        <IconButton
-          aria-label='delete'
-          size='large'
-          target='_blank'
-          href={`http://localhost:3001/files/download/${fileId}`}
-        >
-          <DownloadForOfflineIcon fontSize='inherit' />
-        </IconButton>
-        <Typography fontFamily={'Verdana'} noWrap sx={{ width: 0.6 }}>
-          {fileName}
-        </Typography>
-        <Stack direction='row' justifyContent='flex-end' sx={{ width: 0.4 }}>
-          <DeleteFileModal fileId={fileId} filename={fileName} />
-        </Stack>
+        <DownloadForOfflineIcon fontSize='inherit' />
+      </IconButton>
+      <Typography fontFamily={'Verdana'} noWrap sx={{ width: 0.6 }}>
+        {originalName}
+      </Typography>
+      <Stack direction='row' justifyContent='flex-end' sx={{ width: 0.4 }}>
+        <DeleteFileModal fileName={fileName} originalName={originalName} />
       </Stack>
-    </Box>
-  );
-};
+    </Stack>
+  </Box>
+);
 
 export default SingleFile;
