+++ teccloud-server/pages/files.tsx
@@ -3,17 +3,13 @@ import axios from 'axios';
 import { Fragment, useCallback, useEffect, useState } from 'react';
 import Head from 'next/head';
 import Box from '@mui/material/Box';
-import Button from '@mui/material/Button';
-import Link from '@mui/material/Link';
 import Typography from '@mui/material/Typography';
 import { useDropzone } from 'react-dropzone';
 import Scaffold from '../components/Scaffold';
-import Stack from '@mui/material/Stack';
 import UploadModal from '../components/UploadModal';
 import UploadStatusDialog, {
   UploadStatus,
 } from '../components/UploadStatusDialog';
-import IconButton from '@mui/material/IconButton';
 import SingleFile from '../components/SingleFile';
 
 export const getServerSideProps: GetServerSideUser = async (ctx) => {
@@ -39,7 +35,6 @@ export const getServerSideProps: GetServerSideUser = async (ctx) => {
 
 const Files: AuthenticatedPage = ({ user }) => {
   const [numberDraggedFiles, setNumberDraggedFiles] = useState<number>(0);
-  const [lastFilesDropped, setLastFilesDropped] = useState<any[]>([]);
   const [folderFiles, setFolderFiles] = useState<any[]>([]);
   const [uploadStatus, setUploadStatus] = useState<UploadStatus>({
     status: 'initial',
@@ -59,7 +54,7 @@ const Files: AuthenticatedPage = ({ user }) => {
     };
 
     fetchFiles().catch(console.error);
-  }, []);
+  });
 
   const onUploadProgress = useCallback((e: ProgressEvent) => {
     const percentage = (100 * e.loaded) / e.total;
@@ -77,12 +72,10 @@ const Files: AuthenticatedPage = ({ user }) => {
       acceptedFiles.forEach((file) => formData.append('files', file));
 
       try {
-        const response = await axios.post(
-          'http://localhost:3001/files/upload',
-          formData,
-          { withCredentials: true, onUploadProgress },
-        );
-        setLastFilesDropped(response.data.files);
+        await axios.post('http://localhost:3001/files/upload', formData, {
+          withCredentials: true,
+          onUploadProgress,
+        });
         location.assign('/files');
       } catch (e: any) {
         if (e.response.status === 413) {
@@ -117,7 +110,7 @@ const Files: AuthenticatedPage = ({ user }) => {
       <Head>
         <title>Files - TecCloud</title>
       </Head>
-      <Scaffold user={user}>
+      <Scaffold user={user} folderId={user.folderId}>
         <UploadModal open={isDragActive} numberFiles={numberDraggedFiles} />
         <UploadStatusDialog status={uploadStatus} setStatus={setUploadStatus} />
         <Box
@@ -130,7 +123,7 @@ const Files: AuthenticatedPage = ({ user }) => {
           {folderFiles.length == 0 ? (
             <>
               <Typography paragraph>
-                Oops...you don't have any files yet.
+                Oops...you don&apos;t have any files yet.
               </Typography>
               <input {...getInputProps()} />
               <Typography paragraph>
@@ -138,10 +131,12 @@ const Files: AuthenticatedPage = ({ user }) => {
               </Typography>
             </>
           ) : (
-            folderFiles.map((file, i) => (
-              <div key={i}>
-                <SingleFile fileId={file.fileId} fileName={file.name} />
-              </div>
+            folderFiles.map((file) => (
+              <SingleFile
+                key={file.id}
+                fileName={file.fileName}
+                originalName={file.originalName}
+              />
             ))
           )}
         </Box>
